Feature: PHPCBF code updated
