--- bin/util.js
@@ -1,30 +1,0 @@
-const path = require( 'path' );
-
-/**
- * Return root directory
- *
- * @return {string} root directory
- */
-const getRoot = () => {
-	return path.resolve( __dirname, '../' );
-}
-
-const info = {
-	error: ( message ) => {
-		return `\x1b[31m${ message }\x1b[0m`;
-	},
-	success: ( message ) => {
-		return `\x1b[32m${ message }\x1b[0m`;
-	},
-	warning: ( message ) => {
-		return `\x1b[33m${ message }\x1b[0m`;
-	},
-	message: ( message ) => {
-		return `\x1b[34m${ message }\x1b[0m`;
-	},
-};
-
-module.exports = {
-	getRoot,
-	info,
-};
