+++ package.json
@@ -50,7 +50,7 @@
 		"lint:js:fix": "npm run lint:js -- --fix ./assets/src",
 		"lint:js:report": "npm run lint:js -- --output-file lint-js-report.json --format json .",
 		"lint:php": "vendor/bin/phpcs",
-		"lint:php:fix": "./bin/phpcbf.js",
+		"lint:php:fix": "node ./bin/phpcbf.js",
 		"lint:package-json": "wp-scripts lint-pkg-json --ignorePath .gitignore",
 		"lint:staged": "lint-staged",
 		"prepare": "npm run init",
